Revert "refactor: move seed-sync-cursors to /migrate/api/admin"
This reverts commit 2ff577c2ac3d031454d8e21111b603d7f0acf8d3.
